+++ trio/_core/_run.py
@@ -7,6 +7,8 @@ import select
 import threading
 from collections import deque
 from contextlib import contextmanager, closing
+
+import outcome
 from contextvars import copy_context
 from math import inf
 from time import monotonic
@@ -14,6 +16,7 @@ from time import monotonic
 import attr
 from async_generator import async_generator, yield_, asynccontextmanager
 from sortedcontainers import SortedDict
+from outcome import Error, Value
 
 from . import _public
 from ._entry_queue import EntryQueue, TrioToken
@@ -23,7 +26,6 @@ from ._ki import (
     enable_ki_protection
 )
 from ._multierror import MultiError
-from ._result import Result, Error, Value
 from ._traps import (
     Abort,
     wait_task_rescheduled,
@@ -402,7 +404,7 @@ class Nursery:
             # KeyboardInterrupt), then save that, but still wait until our
             # children finish.
             def aborted(raise_cancel):
-                self._add_exc(Result.capture(raise_cancel).error)
+                self._add_exc(outcome.capture(raise_cancel).error)
                 return Abort.FAILED
 
             self._parent_waiting_in_aexit = True
@@ -542,7 +544,7 @@ class Task:
         # whether we succeeded or failed.
         self._abort_func = None
         if success is Abort.SUCCEEDED:
-            self._runner.reschedule(self, Result.capture(raise_cancel))
+            self._runner.reschedule(self, outcome.capture(raise_cancel))
 
     def _attempt_delivery_of_any_pending_cancel(self):
         if self._abort_func is None:
@@ -609,6 +611,8 @@ class Runner:
 
     entry_queue = attr.ib(default=attr.Factory(EntryQueue))
     trio_token = attr.ib(default=None)
+
+    _NO_SEND = object()
 
     def close(self):
         self.io_manager.close()
@@ -688,9 +692,9 @@ class Runner:
     ################
 
     @_public
-    def reschedule(self, task, next_send=Value(None)):
+    def reschedule(self, task, next_send=_NO_SEND):
         """Reschedule the given task with the given
-        :class:`~trio.hazmat.Result`.
+        :class:`outcome.Outcome`.
 
         See :func:`wait_task_rescheduled` for the gory details.
 
@@ -702,10 +706,13 @@ class Runner:
         Args:
           task (trio.hazmat.Task): the task to be rescheduled. Must be blocked
             in a call to :func:`wait_task_rescheduled`.
-          next_send (trio.hazmat.Result): the value (or error) to return (or
+          next_send (outcome.Outcome): the value (or error) to return (or
             raise) from :func:`wait_task_rescheduled`.
 
         """
+        if next_send is self._NO_SEND:
+            next_send = Value(None)
+
         assert task._runner is self
         assert task._next_send is None
         task._next_send = next_send